assets/css/syntax/README.md.assets/img - delete, should not be deployed
Thisis a work-around the the issue that these files pollute the site
consuming this theme. The README.md even appears in site.pages!

Jekyll's `exclude` configutation does not appear to be able to exclude
these, either in the site or the theme. Puzzled by this, but for the
mean time, I'm just deleting these files in my version of the theme,
pending a better solution.